+++ src/main/java/evaluacioncontinua/service/EmpleadoServiceImpl.java
@@ -1,42 +1,45 @@
 package evaluacioncontinua.service;
 
 import java.util.List;
-
-import org.springframework.beans.factory.annotation.Autowired;
+import evaluacioncontinua.exception.ResourceNotFoundException;
+import evaluacioncontinua.util.Mensajes;
+import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-
 import evaluacioncontinua.dao.IEmpleadoDao;
 import evaluacioncontinua.entity.Empleado;
 
 
 @Service
+@RequiredArgsConstructor
 public class EmpleadoServiceImpl implements IEmpleadoService {
 
-	@Autowired
-	private IEmpleadoDao empleadoDao;
-	
-	
+	private final IEmpleadoDao empleadoDao;
+
 	@Override
 	public List<Empleado> listarEmpleados() {
-	
-		return (List<Empleado>)empleadoDao.findAll();
+		return empleadoDao.findAll();
 	}
 
 	@Override
 	@Transactional
-	public void crearEmpleados(Empleado empleado) {
-		empleadoDao.save(empleado);
+	public Empleado crearEmpleado(Empleado empleado) {
+		return empleadoDao.save(empleado);
 	}
 
 	@Override
+	@Transactional
 	public Empleado editarEmpleado(Long id) {
-		return empleadoDao .findById(id).orElse(null);
+		return empleadoDao.findById(id)
+				.orElseThrow(() -> new ResourceNotFoundException(Mensajes.EMPLEADO_NO_ENCONTRADO + id));
 	}
 
 	@Override
+	@Transactional
 	public void eliminarEmpleado(Long id) {
-		empleadoDao .deleteById(id);
+		if (!empleadoDao.existsById(id)) {
+			throw new ResourceNotFoundException(Mensajes.EMPLEADO_NO_ENCONTRADO + id);
+		}
+		empleadoDao.deleteById(id);
 	}
-
 }
